+++ CHANGES.md
@@ -1,6 +1,7 @@
 #2015-06-24
 
 * Added a new cost difference output
+* Fixed a bug in the counting of traded emissions in the detailed and sectoral emissions charts
 
 #2015-06-23